refactor(desktop): 拆分应用弹窗与伙伴创作辅助模块
